+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Next version
 
+### 💥 Breaking changes
+
+* Removed the option `raise_on_max_attempts` from `Retrier`. If the number of attempts is reached, the retrier will always raise an exception.
+
+### ✨ Improved
+
+* Better typing for `Retrier.__call__()`.
+
 ### 🚀 New
 
 * Add `get_weather` function to retrieve weather data from the LCO API (via `lvmapi`).